more exercises and other things

+++ src/main/java/ar/com/problems/numbers/Greed.java
@@ -14,11 +14,11 @@ public class Greed {
         int score = 0;
         for (Map.Entry<Integer, Long> e : collect.entrySet()) {
             switch (e.getKey()) {
-                case 1: score += ( ( e.getValue() >= 3) ? 1000 : 0) + (e.getValue() % 3) * 100; break;
+                case 1: score += (int) (( ( e.getValue() >= 3) ? 1000 : 0) + (e.getValue() % 3) * 100); break;
                 case 2: score += ( ( e.getValue() >= 3) ? 200 : 0); break;
                 case 3: score += ( ( e.getValue() >= 3) ? 300 : 0); break;
                 case 4: score += ( ( e.getValue() >= 3) ? 400 : 0); break;
-                case 5: score += ( ( e.getValue() >= 3) ? 500 : 0) + (e.getValue() % 3) * 50; break;
+                case 5: score += (int) (( ( e.getValue() >= 3) ? 500 : 0) + (e.getValue() % 3) * 50); break;
                 case 6: score += ( ( e.getValue() >= 3) ? 600 : 0); break;
             }
         }
